hackerrank/balanced_brackets/problem.go
// Copyright 2022 Kurtis Van Gent
//
// Use of this source code is governed by an MIT-style license that can be found
// in the LICENSE file or at https://opensource.org/licenses/MIT.

package balanced_brackets

import (
	"bufio"
	"fmt"
	"io"
	"os"
	"strconv"
	"strings"
)

/*
 * Complete the 'isBalanced' function below.
 *
 * The function is expected to return a STRING.
 * The function accepts STRING s as parameter.
 */

func isBalanced(s string) string {
	q := []rune{}
	for _, r := range s {
		switch r {
		case '{', '(', '[':
			q = append(q, r)
		case '}', ')', ']':
			last := len(q)-1
			if len(q) == 0 || inverse(r) != q[last] {
				return "NO"
			}
			q = q[:last]
		}
	}
	if len(q) > 0 {
		return "NO"
	}
	return "YES"
}

func inverse(r rune) rune {
	switch r {
	case '}':
		return '{'
	case ')':
		return '('
	case ']':
		return '['
	}
	panic(fmt.Sprintf("unexpected rune: %v", r))
}

func main() {
	reader := bufio.NewReaderSize(os.Stdin, 16*1024*1024)

	stdout, err := os.Create(os.Getenv("OUTPUT_PATH"))
	checkError(err)

	defer stdout.Close()

	writer := bufio.NewWriterSize(stdout, 16*1024*1024)

	tTemp, err := strconv.ParseInt(strings.TrimSpace(readLine(reader)), 10, 64)
	checkError(err)
	t := int32(tTemp)

	for tItr := 0; tItr < int(t); tItr++ {
		s := readLine(reader)

		result := isBalanced(s)

		fmt.Fprintf(writer, "%s\n", result)
	}

	writer.Flush()
}

func readLine(reader *bufio.Reader) string {
	str, _, err := reader.ReadLine()
	if err == io.EOF {
		return ""
	}

	return strings.TrimRight(string(str), "\r\n")
}

func checkError(err error) {
	if err != nil {
		panic(err)
	}
}
